requirements.txt
customtkinter==5.1.3
pandas==2.0.2
Pillow==9.5.0
openpyxl==3.1.2